refactor: convert suggested translation contents to use React
This temporarily has an extra class on the directive element, but that
will be removed as soon as the directive encompasses more than just the
translation part of the suggestion panel.
