+++ app/components/SuggestionsBody/index.jsx
@@ -1,8 +1,7 @@
-import Button from '../Button'
 import React from 'react'
 import { IntlMixin } from 'react-intl'
-import SuggestionMatchPercent from '../SuggestionMatchPercent'
-import SuggestionUpdateMessage from '../SuggestionUpdateMessage'
+import SuggestionContents from '../SuggestionContents'
+import SuggestionTranslationDetails from '../SuggestionTranslationDetails'
 
 /**
  * Button that can be disabled.
@@ -23,82 +22,13 @@ let SuggestionsBody = React.createClass({
     })
   },
 
-  getDefaultProps: () => {
-    return {
-      copying: false
-    }
-  },
-
-  /**
-   * Calculate the match type for the suggestion
-   */
-  matchType: function (suggestion) {
-    let topMatch = suggestion.matchDetails[0]
-
-    if (topMatch.type === 'IMPORTED_TM') {
-      return 'imported'
-    }
-    if (topMatch.type === 'LOCAL_PROJECT') {
-      if (topMatch.contentState === 'Translated') {
-        return 'translated'
-      }
-      if (topMatch.contentState === 'Approved') {
-        return 'approved'
-      }
-    }
-    console.error('Unable to generate row display type for top match')
-  },
-
-  user: function (suggestion) {
-    let topMatch = suggestion.matchDetails[0]
-    return topMatch.lastModifiedBy || 'Anonymous'
-  },
-
-  lastChanged: function (suggestion) {
-    let topMatch = suggestion.matchDetails[0]
-    if (topMatch.type === 'IMPORTED_TM') {
-      return topMatch.lastChanged
-    }
-    if (topMatch.type === 'LOCAL_PROJECT') {
-      return topMatch.lastModifiedDate
-    }
-    console.error('match type not recognized for looking up date: ' +
-                  topMatch.type)
-  },
-
   render: function () {
-    let label = this.props.copying ? 'Copying' : 'Copy Translation'
-    let matchType = this.matchType(this.props.suggestion)
-    let user = this.user(this.props.suggestion)
-    let lastChanged = this.lastChanged(this.props.suggestion)
-
     return (
-      <div className="TransUnit-details">
-        <div className="u-floatLeft u-sizeLineHeight-1">
-          <SuggestionUpdateMessage
-            matchType={matchType}
-            user={user}
-            lastChanged={lastChanged}/>
-        </div>
-        <div className="u-floatRight u-sm-floatNone">
-          <ul className="u-listInline u-sizeLineHeight-1">
-            <li>
-              <SuggestionMatchPercent
-                matchType={matchType}
-                percent={this.props.suggestion.similarityPercent}/>
-            </li>
-            <li>
-              <Button
-                className="Button Button--small u-rounded Button--primary
-                           u-sizeWidth-6"
-                disabled={this.props.copying}
-                onClick={this.props.copySuggestion}
-                title={label}>
-                {label}
-              </Button>
-            </li>
-          </ul>
-        </div>
+      <div className="TransUnit-panel TransUnit-translation u-sPV-1-2">
+        <SuggestionContents
+          plural={this.props.suggestion.sourceContents.length > 1}
+          contents={this.props.suggestion.targetContents}/>
+        <SuggestionTranslationDetails {... this.props}/>
       </div>
     )
   }
